Hangman
# Hangman Game

Welcome to my Hangman game repository! This project was born out of my desire to play Hangman with others, and so I decided to create it myself.

## Features

- **Interactive Gameplay**: The game is designed to be interactive and fun. It's not just about guessing letters, it's about the thrill of the game!
- **Sound Effects**: To enhance the gaming experience, a buzzer sound is played whenever a wrong guess is made.
- **Limited Guesses**: The game is challenging! You have a total of 6 guesses to figure out the word.
- **Random Word Selection**: To keep the game fresh and exciting, words are randomly selected based on the category you choose.

This Hangman game is not just a project, it's a testament to my love for coding and creating fun and engaging experiences. Enjoy the game!